moved galleryItem in separate file - implemented update
